Update post detail
